+++ lib/android_pkg/android_pkg.dart
@@ -1,3 +1,5 @@
+import 'package:than_pkg/android_pkg/android_media_util.dart';
+import 'package:than_pkg/android_pkg/android_webview_util.dart';
 import 'package:than_pkg/android_pkg/index.dart';
 
 import 'android_noti_util.dart';
@@ -13,4 +15,6 @@ class AndroidPkg {
   AndroidThumbnail get thumbnail => AndroidThumbnail.thumbnail;
   AndroidWifiUtil get wifi => AndroidWifiUtil.wifi;
   AndroidNotiUtil get notiUtil => AndroidNotiUtil.instance;
+  AndroidWebviewUtil get webview => AndroidWebviewUtil.instance;
+  AndroidMediaUtil get media => AndroidMediaUtil.instance;
 }