+++ lib/linux_pkg/linux_ffmpeg.dart
@@ -1,0 +1,6 @@
+class LinuxFfmpeg {
+   static final LinuxFfmpeg instance = LinuxFfmpeg._();
+  LinuxFfmpeg._();
+  factory LinuxFfmpeg() => instance;
+
+}