+++ lib/linux_pkg/linux_pkg.dart
@@ -1,4 +1,5 @@
 import 'package:than_pkg/linux_pkg/linux_app.dart';
+import 'package:than_pkg/linux_pkg/linux_noti_util.dart';
 import 'package:than_pkg/linux_pkg/linux_thumbnail.dart';
 import 'package:than_pkg/linux_pkg/linux_wifi_util.dart';
 
@@ -10,4 +11,5 @@ class LinuxPkg {
   LinuxThumbnail get thumbnail => LinuxThumbnail.thumbnail;
   LinuxWifiUtil get wifi => LinuxWifiUtil.wifi;
   LinuxApp get app => LinuxApp.app;
+  LinuxNotiUtil get notiUtil => LinuxNotiUtil.instance;
 }